+++ components/BusinessSearchPage.jsx
@@ -65,6 +65,7 @@ var BusinessSearchPage = React.createClass({
         this.setState({displayMobileFilters: (!this.state.displayMobileFilters)});
     },
     handleSearchChange: function (nextSearch) {
+        console.log('ici je passe', nextSearch);
         var search = _.assign({}, this.props.search, nextSearch, { page: 1 });
         this.context.executeAction(BusinessActions.submitSearch, search);
     }
