+++ public/css/style.css
@@ -11838,10 +11838,172 @@ section.home-search .home-search-mobile {
   max-width: 100%;
   overflow-x: hidden;
   width: 100%;
-  font-size: 1em; }
+  font-size: 1em;
+  background-color: white;
+  color: #9a9a9a;
+  font-size: 1.5em; }
+  #content.search .new-filters .filter-header {
+    background-color: #f7f7f7;
+    font-size: 0.8em;
+    height: 60px;
+    text-transform: uppercase;
+    padding-left: 10px;
+    display: -webkit-flexbox;
+    display: -ms-flexbox;
+    display: -webkit-flex;
+    display: flex;
+    -webkit-flex-align: center;
+    -ms-flex-align: center;
+    -webkit-align-items: center;
+    align-items: center;
+    text-align: center;
+    text-align: left;
+    border-bottom: 1px solid #e6e6e6; }
+    #content.search .new-filters .filter-header .header-title {
+      max-width: 60%; }
+  #content.search .new-filters .location, #content.search .new-filters .business-name {
+    padding: 15px 10px 20px 10px;
+    border-bottom: 1px solid #e6e6e6; }
+    #content.search .new-filters .location input, #content.search .new-filters .business-name input {
+      border-top-width: 1px;
+      outline: none;
+      -webkit-appearance: none; }
+    #content.search .new-filters .location .input-group, #content.search .new-filters .business-name .input-group {
+      margin: 10px 0 10px 0;
+      height: 40px; }
+      #content.search .new-filters .location .input-group .form-control, #content.search .new-filters .business-name .input-group .form-control {
+        height: 40px;
+        border-radius: 2px;
+        border-top-right-radius: 0;
+        border-bottom-right-radius: 0;
+        border-color: #e6e6e6;
+        color: #9a9a9a;
+        -webkit-box-shadow: none;
+        box-shadow: none;
+        -webkit-transition: none;
+        -o-transition: none;
+        transition: none;
+        margin-top: 0px; }
+      #content.search .new-filters .location .input-group .input-group-addon, #content.search .new-filters .business-name .input-group .input-group-addon {
+        border-radius: 2px;
+        border-top-left-radius: 0;
+        border-bottom-left-radius: 0;
+        border-color: #e6e6e6; }
+    #content.search .new-filters .location .form-control, #content.search .new-filters .business-name .form-control {
+      height: 40px;
+      border-radius: 2px;
+      border-color: #e6e6e6;
+      color: #9a9a9a;
+      -webkit-box-shadow: none;
+      box-shadow: none;
+      -webkit-transition: none;
+      -o-transition: none;
+      transition: none;
+      margin-top: 10px; }
+    #content.search .new-filters .location .btn-around, #content.search .new-filters .business-name .btn-around {
+      background-color: #f7f7f7;
+      color: #9a9a9a;
+      font-weight: 200;
+      width: 100%;
+      height: 40px;
+      border: 1px solid #e6e6e6;
+      border-radius: 2px; }
+      #content.search .new-filters .location .btn-around:before, #content.search .new-filters .business-name .btn-around:before {
+        content: ' ';
+        background: url(../img/search/location-new.svg) no-repeat center center;
+        background-size: contain;
+        width: 18px;
+        height: 18px;
+        position: absolute;
+        left: calc(50% - 65px); }
   #content.search .new-filters .filters-category {
     display: block;
-    font-size: 1.5em; }
+    height: 50px;
+    border-bottom: 1px solid #e6e6e6;
+    line-height: 50px;
+    padding-left: 10px;
+    padding-right: 10px;
+    color: #9a9a9a; }
+    #content.search .new-filters .filters-category:hover {
+      text-decoration: none; }
+    #content.search .new-filters .filters-category:after {
+      content: '›';
+      float: right;
+      font-family: "Lato", sans-serif;
+      font-size: 1.5em; }
+    #content.search .new-filters .filters-category .tag-count {
+      color: white;
+      font-weight: 400;
+      background-color: #fe5b5f;
+      margin-left: 5px;
+      padding: 1px 5px;
+      border-radius: 3px; }
+  #content.search .new-filters .checkbox-inline {
+    height: 35px;
+    line-height: 35px;
+    font-weight: 200;
+    padding-left: 30px;
+    display: block; }
+  #content.search .new-filters .checkbox-inline + .checkbox-inline {
+    margin-left: 0; }
+  #content.search .new-filters .promo-line {
+    height: 50px;
+    line-height: 50px;
+    border-bottom: 1px solid #e6e6e6; }
+    #content.search .new-filters .promo-line input[type="checkbox"] + span {
+      padding-left: 25px; }
+    #content.search .new-filters .promo-line:before {
+      content: ' ';
+      background: url(../img/search/icon-promo.png) no-repeat center center;
+      background-size: contain;
+      width: 18px;
+      height: 18px;
+      position: absolute;
+      left: 33px;
+      top: 17px; }
+  #content.search .new-filters .filter-footer {
+    position: fixed;
+    bottom: 0;
+    background-color: #f7f7f7;
+    height: 60px;
+    width: 100%;
+    padding: 10px 5px 10px 5px;
+    -webkit-box-shadow: 0px -2px 5px rgba(0, 0, 0, 0.1);
+    -moz-box-shadow: 0px -2px 5px rgba(0, 0, 0, 0.1);
+    box-shadow: 0px -2px 5px rgba(0, 0, 0, 0.1); }
+  #content.search .new-filters .btn-red {
+    margin: 0 5px 0 5px;
+    border-radius: 0;
+    letter-spacing: 0;
+    height: 40px;
+    line-height: 28px;
+    font-size: 1em;
+    font-weight: 200; }
+    #content.search .new-filters .btn-red.semi {
+      width: calc(50% - 10px); }
+    #content.search .new-filters .btn-red.full {
+      margin-left: 0;
+      width: 100%; }
+    #content.search .new-filters .btn-red.previous {
+      position: absolute;
+      top: 10px;
+      right: 10px;
+      display: -webkit-flexbox;
+      display: -ms-flexbox;
+      display: -webkit-flex;
+      display: flex;
+      -webkit-flex-align: center;
+      -ms-flex-align: center;
+      -webkit-align-items: center;
+      align-items: center;
+      text-align: center; }
+      #content.search .new-filters .btn-red.previous:before {
+        content: '<';
+        line-height: 28px;
+        float: left;
+        font-family: "Lato", sans-serif;
+        font-size: 1.5em;
+        margin-right: 5px; }
 
 #content.search .subfilters {
   z-index: 202;
