+++ public/scss/partials/_page-search.scss
@@ -63,9 +63,180 @@
         overflow-x: hidden;
         width: 100%;
         font-size: 1em;
+        background-color: white;
+        color: $grey7;
+        font-size: 1.5em;
+        .filter-header {
+            background-color: $bg4;
+            font-size: 0.8em;
+            height: 60px;
+            text-transform: uppercase;
+            padding-left: 10px;
+            @include flex-center;
+            text-align: left;
+            border-bottom: 1px solid $grey5;
+            .header-title {
+                max-width: 60%;
+            }
+        }
+        .location, .business-name {
+            padding: 15px 10px 20px 10px;
+            border-bottom: 1px solid $grey5;
+            input {
+                border-top-width: 1px;
+                outline: none;
+                -webkit-appearance: none;
+            }
+            .input-group {
+                margin: 10px 0 10px 0;
+                height: 40px;
+                //Location input
+                .form-control {
+                    height: 40px;
+                    border-radius: 2px;
+                    border-top-right-radius: 0;
+                    border-bottom-right-radius: 0;
+                    border-color: $grey5;
+                    color: $grey7;
+                    -webkit-box-shadow: none;
+                    box-shadow: none;
+                    -webkit-transition: none;
+                    -o-transition: none;
+                    transition: none;
+                    margin-top: 0px;
+                }
+                .input-group-addon {
+                    border-radius: 2px;
+                    border-top-left-radius: 0;
+                    border-bottom-left-radius: 0;
+                    border-color: $grey5;
+                }
+            }
+            //Business name input
+            .form-control {
+                height: 40px;
+                border-radius: 2px;
+                border-color: $grey5;
+                color: $grey7;
+                -webkit-box-shadow: none;
+                box-shadow: none;
+                -webkit-transition: none;
+                -o-transition: none;
+                transition: none;
+                margin-top: 10px;
+            }
+            .btn-around {
+                background-color: $bg4;
+                color: $grey7;
+                font-weight: $light;
+                width: 100%;
+                height: 40px;
+                border: 1px solid $grey5;
+                border-radius: 2px;
+                &:before {
+                    content: ' ';
+                    background: url(../img/search/location-new.svg) no-repeat center center;
+                    background-size: contain;
+                    width: 18px;
+                    height: 18px;
+                    position: absolute;
+                    left: calc(50% - 65px);
+                }
+            }
+        }
         .filters-category {
             display: block;
-            font-size: 1.5em;
+            height: 50px;
+            border-bottom: 1px solid $grey5;
+            line-height: 50px;
+            padding-left: 10px;
+            padding-right: 10px;
+            color: $grey7;
+            &:hover {
+                text-decoration: none;
+            }
+            &:after {
+                content: '›';
+                float: right;
+                font-family: $font-second;
+                font-size: 1.5em;
+            }
+            .tag-count {
+                color: white;
+                font-weight: $regular;
+                background-color: $hairfie;
+                margin-left: 5px;
+                padding: 1px 5px;
+                border-radius: 3px;
+            }
+        }
+        .checkbox-inline {
+            height: 35px;
+            line-height: 35px;
+            font-weight: $light;
+            padding-left: 30px;
+            display: block;
+        }
+        .checkbox-inline + .checkbox-inline {
+            margin-left: 0;
+        }
+        .promo-line {
+            height: 50px;
+            line-height: 50px;
+            border-bottom: 1px solid $grey5;
+            input[type="checkbox"] + span {
+                padding-left: 25px;
+            }
+            &:before {
+                content: ' ';
+                background: url(../img/search/icon-promo.png) no-repeat center center;
+                background-size: contain;
+                width: 18px;
+                height: 18px;
+                position: absolute;
+                left: 33px;
+                top: 17px; 
+            }
+        }
+        .filter-footer {
+            position: fixed;
+            bottom: 0;
+            background-color: $bg4;
+            height: 60px;
+            width: 100%;
+            padding: 10px 5px 10px 5px;
+            @include box-shadow (0px -2px 5px rgba(0,0,0,0.1));
+        }
+        .btn-red {
+            margin: 0 5px 0 5px;
+            border-radius: 0;
+            letter-spacing: 0;
+            height: 40px;
+            line-height: 28px;
+            font-size: 1em;
+            font-weight: $light;
+            &.semi {
+                width: calc(50% - 10px);
+            }
+            &.full {
+                margin-left: 0;
+                width: 100%;
+            }
+            &.previous {
+                position: absolute;
+                top: 10px;
+                right: 10px;
+                @include flex-center;
+                &:before {
+                    //content: '‹';
+                    content: '<';
+                    line-height: 28px;
+                    float: left;
+                    font-family: $font-second;
+                    font-size: 1.5em;
+                    margin-right: 5px;
+                }
+            }
         }
     }
     .subfilters {
